Fixed Module Version Checking issue; Fixed issue with Parent Directory when not using Symlink
- Fixed an issue where PSReadline was failing to install the 'beta'
version which is more performant because the FullyQualifiedName
evaluation was failing to install the PreRelease Version from the
Gallery, now a regex splits the Semantic version from the string so the
check doesn't fail.
- Fixed an issue with retrieving the Parent Directory in order to
reference Defaults.yaml and other files, this will now perform a check
to see if the current directory is a Symlink or not.

These updates seem to take the load time from around 7 seconds to 3.5
seconds